feat(ph-ai): make executables ready for plan/research modes
