+++ ee/hogai/chat_agent/prompts/__init__.py
@@ -1,0 +1,40 @@
+from ee.hogai.chat_agent.prompts.base import (
+    AGENT_CORE_MEMORY_PROMPT,
+    AGENT_PROMPT,
+    BASIC_FUNCTIONALITY_PROMPT,
+    CONTEXTUAL_TOOLS_REMINDER_PROMPT,
+    DOING_TASKS_PROMPT,
+    PROACTIVENESS_PROMPT,
+    ROLE_PROMPT,
+    SWITCHING_MODES_PROMPT,
+    TASK_MANAGEMENT_PROMPT,
+    TONE_AND_STYLE_PROMPT,
+    TOOL_USAGE_POLICY_PROMPT,
+    WRITING_STYLE_PROMPT,
+)
+from ee.hogai.core.agent_modes.prompt_builder import (
+    ROOT_BILLING_CONTEXT_ERROR_PROMPT,
+    ROOT_BILLING_CONTEXT_WITH_ACCESS_PROMPT,
+    ROOT_BILLING_CONTEXT_WITH_NO_ACCESS_PROMPT,
+    ROOT_GROUPS_PROMPT,
+)
+
+__all__ = [
+    # Base prompts
+    "AGENT_CORE_MEMORY_PROMPT",
+    "AGENT_PROMPT",
+    "BASIC_FUNCTIONALITY_PROMPT",
+    "CONTEXTUAL_TOOLS_REMINDER_PROMPT",
+    "DOING_TASKS_PROMPT",
+    "PROACTIVENESS_PROMPT",
+    "ROLE_PROMPT",
+    "ROOT_BILLING_CONTEXT_ERROR_PROMPT",
+    "ROOT_BILLING_CONTEXT_WITH_ACCESS_PROMPT",
+    "ROOT_BILLING_CONTEXT_WITH_NO_ACCESS_PROMPT",
+    "ROOT_GROUPS_PROMPT",
+    "SWITCHING_MODES_PROMPT",
+    "TASK_MANAGEMENT_PROMPT",
+    "TONE_AND_STYLE_PROMPT",
+    "TOOL_USAGE_POLICY_PROMPT",
+    "WRITING_STYLE_PROMPT",
+]
