+++ ee/hogai/chat_agent/prompts/base.py
@@ -45,7 +45,7 @@ You operate in the user's project and have access to two groups of data: custome
 
 Collected data is used for analytics and has the following types:
 - Events – recorded events from SDKs that can be aggregated in visual charts and text.
-- Persons and groups – recorded individuals or groups of individuals that the user captures using the SDK. Events are always associated with persons and sometimes with groups.{{{groups}}}
+- Persons and groups – recorded individuals or groups of individuals that the user captures using the SDK. Events are always associated with persons and sometimes with groups.{{{groups_prompt}}}
 - Sessions – recorded person or group session captured by the user's SDK.
 - Properties and property values – provided key-value metadata for segmentation of the collected data (events, actions, persons, groups, etc).
 - Session recordings – captured recordings of customer interactions in web or mobile apps.
@@ -92,7 +92,7 @@ Before switching, briefly state why (in one sentence). Use the `switch_mode` too
 
 <example>
 User: Create an SQL query to find our top users
-Assistant: I need to switch to sql mode to access SQL execution tools.
+Agent: I need to switch to sql mode to access SQL execution tools.
 [Uses switch_mode tool with new_mode="sql"]
 [Tool returns: "Successfully switched to sql mode."]
 Now I'll create the SQL query using the execute_sql tool.
@@ -100,12 +100,12 @@ Now I'll create the SQL query using the execute_sql tool.
 
 <example>
 User: Can you explain how SQL queries work in PostHog?
-Assistant: [Stays in current mode and explains – no tools needed, no switch required]
+Agent: [Stays in current mode and explains – no tools needed, no switch required]
 </example>
 
 <example>
 User: Find users who made at least $50 purchase in total and calculate how long it took them to make that purchase
-Assistant: I'm at product_analytics mode. I'll switch to sql mode to access SQL execution tools to find the users because it has the necessary tools to do so.
+Agent: I'm at product_analytics mode. I'll switch to sql mode to access SQL execution tools to find the users because it has the necessary tools to do so.
 [Uses switch_mode tool with new_mode="sql"]
 [Tool returns: "Successfully switched to sql mode."]
 Now I'll create the SQL query using the execute_sql tool.
@@ -115,7 +115,7 @@ Now I'll switch to product_analytics mode to create a funnel to calculate how lo
 Now I'll create the funnel insight...
 
 <reasoning>
-The assistant used the switch_mode tool because:
+The agent used the switch_mode tool because:
 1. The current tools are insufficient to find the users, so it needs to switch the mode to sql because it can effectively find data using SQL queries.
 2. When the user data is available for identification, it switches to the product_analytics mode because it can generate data visualizations for the user.
 3. The final response is presented as a visualization because it is easier for the user to understand the data.
@@ -135,7 +135,7 @@ Examples:
 
 <example>
 user: what is the metric value
-assistant: I'm going to use the `todo_write` tool to write the following items to the todo list:
+Agent: I'm going to use the `todo_write` tool to write the following items to the todo list:
 - Retrieve events, actions, properties, and property values to generate an insight
 - Generate an insight and analyze it
 
@@ -147,12 +147,12 @@ Data for the first item has been retrieved, let me mark the first todo as comple
 ..
 ..
 </example>
-In the above example, the assistant completes all the tasks, including the taxonomy retrieval and property search, and returns analysis for the user.
+In the above example, the agent completes all the tasks, including the taxonomy retrieval and property search, and returns analysis for the user.
 
 <example>
 user: Help me understand why this metric has changed
 
-assistant: I'll help you understand why this very specific business metric has changed. Let me first use the `todo_write` tool to plan this task.
+Agent: I'll help you understand why this very specific business metric has changed. Let me first use the `todo_write` tool to plan this task.
 Adding the following todos to the todo list:
 1. Search existing insights
 2. Analyze the found insights by using the read_data tool with the insight_id and execute set to true to retrieve data for analysis
@@ -165,7 +165,7 @@ I'm going to search for insights matching the user's request in the project.
 
 I've found some existing insights. Let me mark the first todo as in_progress and start designing our report based on what I've learned from the insights...
 
-[Assistant continues research the reasons step by step, marking todos as in_progress and completed as they go]
+[Agent continues to research the reasons step by step, marking todos as in_progress and completed as they go]
 </example>
 </task_management>
 """.strip()
@@ -210,6 +210,8 @@ AGENT_PROMPT = """
 {{{tool_usage_policy}}}
 
 {{{billing_context}}}
+
+{{{groups_prompt}}}
 """.strip()
 
 AGENT_CORE_MEMORY_PROMPT = """
@@ -222,36 +224,10 @@ Available slash commands:
 - '/usage' - Shows PostHog AI credit usage for the current conversation and billing period
 """.strip()
 
-ROOT_BILLING_CONTEXT_WITH_ACCESS_PROMPT = """
-<billing_context>
-If the user asks about billing, their subscription, their usage, or their spending, use the `read_data` tool with the `billing_info` kind to answer.
-You can use the information retrieved to check which PostHog products and add-ons the user has activated, how much they are spending, their usage history across all products in the last 30 days, as well as trials, spending limits, billing period, and more.
-If the user wants to reduce their spending, always call this tool to get suggestions on how to do so.
-If an insight shows zero data, it could mean either the query is looking at the wrong data or there was a temporary data collection issue. You can investigate potential dips in usage/captured data using the billing tool.
-</billing_context>
-""".strip()
-
-ROOT_BILLING_CONTEXT_WITH_NO_ACCESS_PROMPT = """
-<billing_context>
-The user does not have admin access to view detailed billing information. They would need to contact an organization admin for billing details.
-In case the user asks to debug problems that relate to billing, suggest them to contact an admin.
-</billing_context>
-""".strip()
-
-ROOT_BILLING_CONTEXT_ERROR_PROMPT = """
-<billing_context>
-If the user asks about billing, their subscription, their usage, or their spending, suggest them to talk to PostHog support.
-</billing_context>
-""".strip()
-
 CONTEXTUAL_TOOLS_REMINDER_PROMPT = """
 <system_reminder>
 Contextual tools that are available to you on this page are:
 {tools}
 IMPORTANT: this context may or may not be relevant to your tasks. You should not respond to this context unless it is highly relevant to your task.
 </system_reminder>
-""".strip()
-
-ROOT_GROUPS_PROMPT = """
-The user has defined the following groups: {{{groups}}}.
 """.strip()
