+++ ee/hogai/core/agent_modes/presets/test/test_sql_preset.py
@@ -57,6 +57,7 @@ class TestSQLAgentNode(BaseTest):
                 ],
             ),
         ):
+            state_1 = AssistantState(messages=[HumanMessage(content="execute sql query")], agent_mode=AgentMode.SQL)
             context_manager = AssistantContextManager(
                 team=self.team, user=self.user, config=RunnableConfig(configurable={})
             )
@@ -67,11 +68,10 @@ class TestSQLAgentNode(BaseTest):
                     NodePath(name=AssistantNodeName.ROOT, message_id="test_id", tool_call_id="test_tool_call_id"),
                 ),
                 context_manager=context_manager,
-                mode=AgentMode.SQL,
+                state=state_1,
             )
             node = mode_manager.node
 
-            state_1 = AssistantState(messages=[HumanMessage(content="execute sql query")])
             next_state = await node.arun(state_1, {})
             assert isinstance(next_state, PartialAssistantState)
             # The state includes context messages + original message + generated message
