+++ ee/hogai/utils/feature_flags.py
@@ -3,16 +3,6 @@ import posthoganalytics
 from posthog.models import Team, User
 
 
-def has_create_form_tool_feature_flag(team: Team, user: User) -> bool:
-    return posthoganalytics.feature_enabled(
-        "phai-create-form-tool",
-        str(user.distinct_id),
-        groups={"organization": str(team.organization_id)},
-        group_properties={"organization": {"id": str(team.organization_id)}},
-        send_feature_flag_events=False,
-    )
-
-
 def has_web_search_feature_flag(team: Team, user: User) -> bool:
     return posthoganalytics.feature_enabled(
         "phai-web-search",
@@ -69,16 +59,6 @@ def has_upsert_dashboard_feature_flag(team: Team, user: User) -> bool:
 def has_error_tracking_mode_feature_flag(team: Team, user: User) -> bool:
     return posthoganalytics.feature_enabled(
         "posthog-ai-error-tracking-mode",
-        str(user.distinct_id),
-        groups={"organization": str(team.organization_id)},
-        group_properties={"organization": {"id": str(team.organization_id)}},
-        send_feature_flag_events=False,
-    )
-
-
-def has_create_notebook_tool_feature_flag(team: Team, user: User) -> bool:
-    return posthoganalytics.feature_enabled(
-        "phai-create-notebook-tool",
         str(user.distinct_id),
         groups={"organization": str(team.organization_id)},
         group_properties={"organization": {"id": str(team.organization_id)}},
